panoptik/requirements.txt
pefile
python-magic
capstone
iocextract
validators
boto3